Move more state tree logic into the Arbor class
