+++ packages/arbor-store/src/MutationEngine.ts
@@ -1,42 +1,6 @@
 import { Arbor } from "./Arbor"
 import { Path } from "./Path"
-import { Seed } from "./Seed"
-import { Mutation, MutationResult, Node } from "./types"
-
-type Clonnable<T extends object> = {
-  $clone(): T
-}
-
-function hasCustomClonningLogic<T extends object>(
-  value: object
-): value is Clonnable<T> {
-  return "$clone" in value && typeof value.$clone === "function"
-}
-
-function cloneViaConstructor<T extends object>(value: T): T {
-  const constructor = value.constructor as new () => T
-  const instance = new constructor()
-  return Object.assign(instance, value)
-}
-
-/**
- * Shallow clone the given value by following a few rules:
- *
- * 1. when value defines a $clone function method, that is used to determine
- * the clonned value;
- * 2. otherwise, the clonned value happens by:
- *    1. creating a new instance via the value's constructor;
- *    2. assign all the properties from the original value to the new instance.
- */
-function clone<T extends object>(value: T): T {
-  const clonedValue = hasCustomClonningLogic<T>(value)
-    ? value.$clone()
-    : cloneViaConstructor(value)
-
-  const seed = Seed.from(value)
-  Seed.plant(clonedValue, seed)
-  return clonedValue
-}
+import { Mutation, MutationResult } from "./types"
 
 export type MutationMode = "eager" | "snapshot"
 
@@ -46,24 +10,15 @@ export class MutationEngine<T extends object> {
     private readonly mode: MutationMode = "eager"
   ) {}
 
-  clone<V extends object>(node: Node<V>): Node<V> {
-    return this.tree.createNode<V>(
-      this.tree.getPathFor(node),
-      this.mode === "snapshot" ? clone(node.$value) : node.$value,
-      this.tree.getLinkFor(node),
-      node.$subscriptions
-    )
-  }
-
   mutate<V extends object>(
     path: Path,
     mutation: Mutation<V>
   ): MutationResult<T> {
     try {
-      const root = this.clone(this.tree.root)
+      const root = this.tree.cloneNode(this.tree.root)
 
       const targetNode = path.walk<V>(root, (child, parent) => {
-        const childCopy = this.clone(child)
+        const childCopy = this.tree.cloneNode(child)
         const link = this.tree.getLinkFor(childCopy)
 
         // TODO: remove mode, it seems we don't necessarily need to employ
