+++ packages/arbor-store/src/guards.ts
@@ -1,6 +1,6 @@
 import { Arbor } from "./Arbor"
 import { ArborProxiable } from "./decorators"
-import type { ArborNode, Node } from "./types"
+import type { ArborNode, Clonnable, Node } from "./types"
 
 export function isNode<T extends object>(value: unknown): value is Node<T> {
   const isNodeValue = value as Node<T>
@@ -22,4 +22,10 @@ export function isProxiable(value: unknown): value is object {
     value[ArborProxiable] === true ||
     value.constructor === Object
   )
+}
+
+export function hasCustomClonningLogic<T extends object>(
+  value: object
+): value is Clonnable<T> {
+  return "$clone" in value && typeof value.$clone === "function"
 }
