+++ packages/arbor-store/src/types.ts
@@ -86,3 +86,7 @@ export type MutationResult<T extends object> = {
   root: Node<T>
   metadata: MutationMetadata
 }
+
+export type Clonnable<T extends object> = {
+  $clone(): T
+}
